+++ api/src/service/submission/index.ts
@@ -1,3 +1,4 @@
+import { SubmissionDeleteService } from './submission.delete.service'
 import { SubmissionHookService } from './submission.hook.service'
 import { SubmissionNotificationService } from './submission.notification.service'
 import { SubmissionService } from './submission.service'
@@ -7,6 +8,7 @@ import { SubmissionStatisticService } from './submission.statistic.service'
 import { SubmissionTokenService } from './submission.token.service'
 
 export const submissionServices = [
+  SubmissionDeleteService,
   SubmissionHookService,
   SubmissionNotificationService,
   SubmissionService,